Move invocation builder inside invocation class
