+++ subprojects/core-exec/src/main/java/dev/nokee/core/exec/CommandLine.java
@@ -55,12 +55,12 @@ public interface CommandLine {
 	CommandLineToolArguments getArguments();
 
 	/**
-	 * Prepares a new invocation via the {@link CommandLineToolInvocationBuilder}.
+	 * Prepares a new invocation via the {@link CommandLineToolInvocation.Builder}.
 	 * The invocation is responsible for configuring the working directory, environment variables, standard stream manipulation, etc.
 	 *
-	 * @return a {@link CommandLineToolInvocationBuilder} instance, never null.
+	 * @return a {@link CommandLineToolInvocation.Builder} instance, never null.
 	 */
-	CommandLineToolInvocationBuilder newInvocation();
+	CommandLineToolInvocation.Builder newInvocation();
 
 	/**
 	 * Convenience for {@code newInvocation().build().submit(engine)}.
